Refactor - code repeated in SM readers is moved to smPreferenceInstance This commit is made on the basic principle of not repeating oneself.
