+++ src/algmatch/stableMatchings/stableMarriageProblem/fileReader.py
@@ -1,7 +1,6 @@
 """
 Class to read in a file of preferences for the Stable Marriage Problem stable matching algorithm.
 """
-from itertools import product
 
 from algmatch.abstractClasses.abstractReader import AbstractReader
 from algmatch.abstractClasses.abstractReader import ReaderError
@@ -12,40 +11,7 @@ class FileReader(AbstractReader):
         super().__init__(filename)
         self._read_data()
 
-    def check_preference_lists(self) -> None:
-        for m, m_prefs in self.men.items():
-
-            if len(set(m_prefs["list"])) != len(m_prefs["list"]):
-                raise ReaderError(f"man {m}", "Repetition in preference list.")
-            
-            for w in m_prefs["list"]:
-                if w not in self.women:
-                    raise ReaderError(f"man {m}", f"Woman {w} not instantiated.")
-            
-        for w, w_prefs in self.women.items():
-
-            if len(set(w_prefs["list"])) != len(w_prefs["list"]):
-                raise ReaderError(f"woman {w}", "Repetition in preference list.")
-            
-            for m in w_prefs["list"]:
-                if m not in self.men:
-                    raise ReaderError(f"woman {w}", f"Man {m} not instantiated.")
-
-    def clean_unacceptable_pairs(self) -> None:
-        for m, w in product(self.men, self.women):
-            if m not in self.women[w]["list"] or w not in self.men[m]["list"]:
-                try: m.remove(w)
-                except ValueError: pass
-                try: w.remove(m)
-                except ValueError: pass
-
-    def set_up_rankings(self):
-        for m in self.men:
-            self.men[m]["rank"] = {woman: idx for idx, woman in enumerate(self.men[m]["list"])}
-        for w in self.women:
-            self.women[w]["rank"] = {man: idx for idx, man in enumerate(self.women[w]["list"])}
-    
-    def parse_file(self) -> None:
+    def _read_data(self) -> None:
         self.no_men = 0
         self.no_women = 0
         self.men = {}                
@@ -95,9 +61,3 @@ class FileReader(AbstractReader):
             preferences = [f"m{i}" for i in entry[1:]]
 
             self.women[woman] = {"list": preferences, "rank": {}}
-    
-    def _read_data(self) -> None:
-        self.parse_file()
-        self.check_preference_lists()
-        self.clean_unacceptable_pairs()
-        self.set_up_rankings()
